Add a schema to the database (#38)
This is a major change.

**breaking changes**
- Python's API has drastically changed.
- Collections have been removed.
- XTR support has been removed temporarily.

We introduce a schema to the database. We can index/store/filter by
different data types, and we can compose different queries and ways to
score.

# What problem does this solve?
ColBERT and more heavyweight retrieval mechanisms can be slow, because
there are more embeddings to compare per document. This makes it
necessary to filter documents or iteratively reduce the amount of
documents scored.

# How did we solve it?
Schemas enable more flexible queries. Filtering becomes an option, and
we can choose to score documents based on each matched element.

## DocumentProcessor
Our main new abstraction is document processing. This has been broken
out from index writing. The DocProcessor branches for each data type
supported, and we optionally quantize tensors as part of this.

ColBERT fields are a special case. ColBERT is both indexed and
contextual, in that we search the index but don't retrieve data from
that field. During scoring, we scan the context field to get all token
embeddings at once.

## Scoring
Retrievers have been generalized into scoring. It's still a WIP, but we
have the concept of retrieval and ranking. Combined with different types
of fields, we can think of ColBERT as indexed with contextual data and
XTR as indexed only.

## Collections
Collections have been removed. Collections enabled an easier way to
index data by passing text and automatically embedding it in LintDB.
However, this conflates the main idea behind LintDB -- storing and
retrieving. We see collections coming back as extensions within the
Python library.

# Python bindings
Python bindings were using SWIG. SWIG files used a custom syntax to
define what C++ was bound to Python. This became difficult to maintain,
because some of our data objects made sense to be translated to Python
dictionaries. This wasn't simple to accomplish.

We've migrated to nanobind instead of SWIG. Nanobind is declared in C++
directly. There are still some growing pains with this, but it's much
clearer how to define, override, or rename our bindings.

# Documentation
Documentation is moving to mkdocs instead of sphinx. The main problem
here was versioning our documentation. Sphinx did not have a clear
enough way to handle this automatically. Mkdocs, however, has mike to
version docs.

We haven't figured out all of the bugs with translating our docstrings,
but fixing this seems doable.

+++ lintdb/index.h
@@ -1,23 +1,25 @@
 #ifndef LINTDB_INDEX_IVF_H
 #define LINTDB_INDEX_IVF_H
 
-#include <stdint.h>
+#include <cstdint>
 #include <memory>
 #include <string>
 #include <unordered_map>
 #include <vector>
 
 #include <unordered_set>
-#include "lintdb/EmbeddingBlock.h"
-#include "lintdb/Encoder.h"
-#include "lintdb/Passages.h"
-#include "lintdb/SearchOptions.h"
-#include "lintdb/SearchResult.h"
 #include "lintdb/api.h"
 #include "lintdb/exception.h"
+#include "lintdb/invlists/IndexWriter.h"
 #include "lintdb/invlists/InvertedList.h"
-#include "lintdb/retrievers/PlaidRetriever.h"
-#include "lintdb/retrievers/Retriever.h"
+#include "lintdb/quantizers/CoarseQuantizer.h"
+#include "lintdb/query/Query.h"
+#include "lintdb/schema/DocProcessor.h"
+#include "lintdb/schema/Document.h"
+#include "lintdb/schema/FieldMapper.h"
+#include "lintdb/schema/Schema.h"
+#include "lintdb/SearchOptions.h"
+#include "lintdb/SearchResult.h"
 #include "lintdb/version.h"
 
 // forward declare these classes and avoid including the rocksdb headers.
@@ -39,20 +41,9 @@ struct Configuration {
     Version lintdb_version =
             LINTDB_VERSION; /// the current version of the index. Used
                             /// internally for feature compatibility.
-    size_t nlist = 256;     /// the number of centroids to train.
-    size_t nbits = 2;  /// the number of bits to use in residual compression.
-    size_t niter = 10; /// the number of iterations to use during training.
-    size_t dim;        /// the dimensions expected for incoming vectors.
-    size_t num_subquantizers =
-            16; /// the number of subquantizers to use in the product quantizer.
-    IndexEncoding quantizer_type =
-            IndexEncoding::BINARIZER; /// whether to compress residuals.
 
     inline bool operator==(const Configuration& other) const {
-        return nlist == other.nlist && nbits == other.nbits &&
-                niter == other.niter && dim == other.dim &&
-                quantizer_type == other.quantizer_type &&
-                num_subquantizers == other.num_subquantizers;
+        return lintdb_version == other.lintdb_version;
     }
 
     Configuration() = default;
@@ -75,20 +66,11 @@ struct IndexIVF {
     /// load an existing index.
     IndexIVF(const std::string& path, bool read_only = false);
 
-    IndexIVF(std::string path, Configuration& config);
-
     IndexIVF(
-            std::string path, /// path to the database.
-            size_t nlist,     /// number of centroids to use in L1 quantizing.
-            size_t dim,       /// number of dimensions per embedding.
-            size_t binarize_nbits =
-                    2, /// nbits used in the LSH encoding for esiduals.
-            size_t niter = 10,
-            size_t num_subquantizers = 16,
-            IndexEncoding quantizer_type = IndexEncoding::BINARIZER,
-            bool read_only = false);
+            const std::string& path,
+            const Schema& schema,
+            const Configuration& config);
 
-    // TODO(mbarta): this breaks SWIG. SWIG needs to ignore this, but won't.
     /**
      * Copy creates a new index at the given path from a trained index. The copy
      * will always be writeable.
@@ -98,53 +80,21 @@ struct IndexIVF {
      *
      * @param path the path to initialize the index.
      */
-    IndexIVF(const IndexIVF& other, const std::string path);
+    IndexIVF(const IndexIVF& other, const std::string& path);
 
     /**
      * Train will learn quantization and compression parameters from the given
      * data.
      *
-     * @param n the number of embeddings to train on.
-     * @param embeddings the embeddings to train on.
-     * @param nlist the number of centroids to train. Required if not provided
-     * at initialization.
-     * @param niter the number of iterations to train. Required if not provided
-     * at initialization.
      */
-    void train(
-            size_t n,
-            std::vector<float>& embeddings,
-            size_t nlist = 0,
-            size_t niter = 0);
-    void train(
-            float* embeddings,
-            size_t n,
-            size_t dim,
-            size_t nlist = 0,
-            size_t niter = 0);
-    void train(
-            float* embeddings,
-            int n,
-            int dim,
-            size_t nlist = 0,
-            size_t niter = 0);
+    void train(const std::vector<Document>& docs);
 
-    /**
-     * set_centroids overwrites the centroids in the encoder.
-     *
-     * This is useful if you want to parallelize index writing and merge indices
-     * later.
-     */
-    void set_centroids(float* data, int n, int dim);
-
-    /**
-     * set_weights overwrites the compression weights in the encoder, if using
-     * compression.
-     */
-    void set_weights(
-            const std::vector<float> weights,
-            const std::vector<float> cutoffs,
-            const float avg_residual);
+    void set_quantizer(
+            const std::string& field,
+            std::shared_ptr<Quantizer> quantizer);
+    void set_coarse_quantizer(
+            const std::string& field,
+            std::shared_ptr<ICoarseQuantizer> quantizer);
 
     /**
      * search will find the nearest neighbors for a vector block.
@@ -157,31 +107,7 @@ struct IndexIVF {
      */
     std::vector<SearchResult> search(
             const uint64_t tenant,
-            const float* data,
-            const int n,
-            const int dim,
-            const size_t k,
-            const SearchOptions& opts = SearchOptions()) const;
-
-    std::vector<SearchResult> search(
-            const uint64_t tenant,
-            const EmbeddingBlock& block,
-            const size_t k,
-            const SearchOptions& opts = SearchOptions()) const;
-
-    std::vector<SearchResult> search(
-            const uint64_t tenant,
-            const float* data,
-            const int n,
-            const int dim,
-            const size_t n_probe,
-            const size_t k,
-            const SearchOptions& opts = SearchOptions()) const;
-
-    std::vector<SearchResult> search(
-            const uint64_t tenant,
-            const EmbeddingBlock& block,
-            const size_t n_probe,
+            const Query& query,
             const size_t k,
             const SearchOptions& opts = SearchOptions()) const;
 
@@ -192,12 +118,12 @@ struct IndexIVF {
      * @param docs a vector of EmbeddingPassages. This includes embeddings and
      * ids.
      */
-    void add(const uint64_t tenant, const std::vector<EmbeddingPassage>& docs);
+    void add(const uint64_t tenant, const std::vector<Document>& docs);
 
     /**
      * Add a single document.
      */
-    void add_single(const uint64_t tenant, const EmbeddingPassage& doc);
+    void add_single(const uint64_t tenant, const Document& doc);
 
     /**
      * Remove deletes documents from the index by id.
@@ -210,9 +136,7 @@ struct IndexIVF {
     /**
      * Update is a convenience function for remove and add.
      */
-    void update(
-            const uint64_t tenant,
-            const std::vector<EmbeddingPassage>& docs);
+    void update(const uint64_t tenant, const std::vector<Document>& docs);
 
     /**
      * Merge will combine the index with another index.
@@ -223,7 +147,7 @@ struct IndexIVF {
      * multiprocess building of indices.
      *
      */
-    void merge(const std::string path);
+    void merge(const std::string& path);
 
     /**
      * Index should be able to resume from a previous state.
@@ -236,8 +160,11 @@ struct IndexIVF {
     void close();
 
     ~IndexIVF() {
-        for (auto cf : column_families) {
-            db->DestroyColumnFamilyHandle(cf);
+        for (auto& cf : column_families) {
+            if (cf) {
+                auto status = db->DestroyColumnFamilyHandle(cf);
+                assert(status.ok());
+            }
         }
     }
 
@@ -245,21 +172,27 @@ struct IndexIVF {
     std::string path;
     std::shared_ptr<rocksdb::DB> db;
     std::vector<rocksdb::ColumnFamilyHandle*> column_families;
+    Schema schema;
+    std::shared_ptr<FieldMapper> field_mapper;
 
-    std::shared_ptr<Encoder> encoder;
-    std::shared_ptr<Quantizer> quantizer;
-    std::unique_ptr<Retriever> retriever;
+    std::unordered_map<std::string, std::shared_ptr<ICoarseQuantizer>>
+            coarse_quantizer_map;
+    std::unordered_map<std::string, std::shared_ptr<Quantizer>> quantizer_map;
 
-    // helper to initialize the inverted list.
-    void initialize_inverted_list(Version& version);
-    // helper to initialize the encoder, quantizer, and retrievers. These are
-    // all inter-related.
-    void initialize_retrieval(IndexEncoding quantizer_type);
-    // instead of initializing, load from disk.
-    void load_retrieval(std::string path, const Configuration& config);
-
+    std::shared_ptr<DocumentProcessor> document_processor;
+    // Note: invertedList and ForwardIndex are becoming read-only classes for
+    // retrieval. writing is done through the index writer. Merging/Removing
+    // will likely move to the writer as well.
     std::shared_ptr<InvertedList> inverted_list_;
     std::shared_ptr<ForwardIndex> index_;
+
+    // helper to initialize the inverted list.
+    void initialize_inverted_list(const Version& version);
+    // helper to initialize the encoder, quantizer, and retrievers. These are
+    // all inter-related.
+    void initialize_retrieval();
+    // instead of initializing, load from disk.
+    void load_retrieval(const std::string& path, const Configuration& config);
 
     /**
      * Flush data to disk.
@@ -274,7 +207,7 @@ struct IndexIVF {
      *
      */
     void write_metadata();
-    Configuration read_metadata(std::string path);
+    Configuration read_metadata(const std::string& path);
 };
 
 } // namespace lintdb
